# docs/context/api/openapi.yaml
openapi: 3.1.0
info:
  title: My-ERP API
  version: 0.1.0
  description: |
    Canonical API contract for My-ERP (finance v1).
    After edits run:
      node .ai/skills/features/context-awareness/scripts/ctl-context.mjs touch
      node .ai/scripts/ctl-api-index.mjs generate --touch
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /v1/ledger/trial-balance:
    get:
      operationId: getTrialBalance
      summary: Trial balance (试算平衡表) — derived from posted vouchers
      tags: [ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: period
          in: query
          required: false
          schema: { type: string, pattern: '^\d{4}-\d{2}$' }
          description: Accounting period (YYYY-MM). Prior movement is rolled into opening balance.
      responses:
        '200':
          description: Trial balance
          content: { application/json: { schema: { $ref: '#/components/schemas/TrialBalance' } } }
        '401': { description: Missing or invalid bearer token }
        '403': { description: Ledger not in org / lacks read }
  /v1/ledger/accounts/{code}:
    get:
      operationId: getAccountLedger
      summary: Subsidiary ledger (明细分类账) for one account — derived, running balance
      tags: [ledger]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
        - name: period
          in: query
          required: false
          schema: { type: string, pattern: '^\d{4}-\d{2}$' }
          description: Accounting period (YYYY-MM). Prior movement is rolled into opening balance.
      responses:
        '200':
          description: Account ledger
          content: { application/json: { schema: { $ref: '#/components/schemas/AccountLedger' } } }
        '401': { description: Missing or invalid bearer token }
        '403': { description: Ledger not in org / lacks read }
...
  /v1/payments:
    get:
      operationId: listPayments
      summary: List 出纳收付款单 (PaymentDoc)
      tags: [payments]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, required: false, schema: { type: string, enum: [draft, pending_approval, approved, confirmed, void] } }
        - { name: direction, in: query, required: false, schema: { type: string, enum: [receipt, payment] } }
      responses:
        '200': { description: Payment docs, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/PaymentDoc' } } } } }
    post:
      operationId: createPayment
      summary: Create a 收/付款单 draft
      tags: [payments]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreatePayment' } } }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentDoc' } } } }
        '400': { description: Invalid input / closed period / bad account }
  /v1/payments/{id}:
    get:
      operationId: getPayment
      summary: Get a payment doc
      tags: [payments]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: Payment doc, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentDoc' } } } }
        '404': { description: Not found }
  /v1/payments/{id}/submit:
    post:
      operationId: submitPayment
      summary: draft → pending_approval (opens the approver task)
      tags: [payments]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentActionRequest' } } } }
      responses:
        '200': { description: Submitted, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentDoc' } } } }
        '400': { description: Wrong status / closed period }
        '409': { description: Stale version }
  /v1/payments/{id}/approve:
    post:
      operationId: approvePayment
      summary: pending_approval → approved (SoD; opens the cashier task)
      tags: [payments]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentActionRequest' } } } }
      responses:
        '200': { description: Approved, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentDoc' } } } }
        '400': { description: Wrong status }
        '403': { description: Approver is the maker (SoD) }
        '409': { description: Stale version }
  /v1/payments/{id}/confirm:
    post:
      operationId: confirmPayment
      summary: approved → confirmed — generate + post the settlement voucher
      tags: [payments]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentConfirmRequest' } } } }
      responses:
        '200': { description: Confirmed, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentConfirmResult' } } } }
        '400': { description: Wrong status / closed period }
        '403': { description: Confirmer is the maker (SoD) }
        '409': { description: Stale version }
  /v1/payments/{id}/void:
    post:
      operationId: voidPayment
      summary: 作废 a pre-confirmed payment doc
      tags: [payments]
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentVoidRequest' } } } }
      responses:
        '200': { description: Voided, content: { application/json: { schema: { $ref: '#/components/schemas/PaymentDoc' } } } }
        '400': { description: Already confirmed / void }
        '409': { description: Stale version }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS256 dev token (P0b mock identity). Carries sub/orgId/ledgerBookId
        (+ optional email). Roles are resolved from Membership, not the token.
        Real Logto (OIDC/JWKS) replaces the verifier later.
  schemas:
...
    VoucherLine:
      type: object
      required: [accountCode, summary]
      properties:
        accountCode: { type: string }
        accountName: { type: string, description: 'Denormalized from the chart (server-set)' }
        summary: { type: string }
        debit: { type: string, nullable: true, description: '2dp; exactly one of debit/credit' }
        credit: { type: string, nullable: true }
        cashFlowItem: { type: string, nullable: true }
    FastEntryDraftPayload:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version: { type: integer, enum: [1] }
        summary: { type: string, maxLength: 500 }
        contractId: { type: string, format: uuid, nullable: true }
        lines:
          type: array
          maxItems: 100
          items:
            type: object
            additionalProperties: false
            properties:
              accountCode: { type: string, maxLength: 64 }
              accountName: { type: string, maxLength: 120 }
              summary: { type: string, maxLength: 500 }
              debit: { type: string, description: '2dp draft amount, optional' }
              credit: { type: string, description: '2dp draft amount, optional' }
              cashFlowItem: { type: string, maxLength: 64 }
    Voucher:
      type: object
      required: [id, ledgerBookId, no, date, period, status, summary, totalDebit, totalCredit, maker, lines]
      properties:
        id: { type: string, format: uuid }
        ledgerBookId: { type: string, format: uuid }
        no: { type: string }
        date: { type: string, format: date }
        period: { type: string, example: '2026-06' }
        status: { type: string, enum: [draft, pending, posted, reversed] }
        summary: { type: string }
        totalDebit: { type: string }
        totalCredit: { type: string }
        maker: { type: string }
        checker: { type: string, nullable: true }
        postedAt: { type: string, format: date-time, nullable: true }
        reversalOf: { type: string, format: uuid, nullable: true }
        reversedBy: { type: string, format: uuid, nullable: true }
        attachments: { type: integer }
        contractId: { type: string, format: uuid, nullable: true, description: 'Contract link (T-005)' }
        draftPayload:
          allOf:
            - $ref: '#/components/schemas/FastEntryDraftPayload'
          nullable: true
        lines:
          type: array
          items: { $ref: '#/components/schemas/VoucherLine' }
...
    PaymentDoc:
      type: object
      required: [id, no, direction, date, period, counterparty, summary, amount, cashAccountCode, contraAccountCode, status, maker, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        no: { type: string, example: 收-2026-06-001 }
        direction: { type: string, enum: [receipt, payment] }
        date: { type: string, format: date }
        period: { type: string, example: 2026-06 }
        counterparty: { type: string }
        summary: { type: string }
        amount: { type: string, description: '2dp' }
        cashAccountCode: { type: string }
        contraAccountCode: { type: string }
        status: { type: string, enum: [draft, pending_approval, approved, confirmed, void] }
        settlementVoucherId: { type: string, format: uuid, nullable: true }
        contractId: { type: string, format: uuid, nullable: true, description: 'Reserved for T-005' }
        maker: { type: string }
        approver: { type: string, nullable: true }
        confirmer: { type: string, nullable: true }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreatePayment:
      type: object
      required: [direction, date, counterparty, summary, amount, cashAccountCode, contraAccountCode]
      properties:
        direction: { type: string, enum: [receipt, payment] }
        date: { type: string, format: date }
        counterparty: { type: string }
        summary: { type: string }
        amount: { type: string, description: 'positive, ≤2dp' }
        cashAccountCode: { type: string, description: '货币资金类 (1001/1002/1012)' }
        contraAccountCode: { type: string, description: '对方科目' }
        contractId: { type: string, format: uuid, nullable: true, description: 'Optional Contract link (T-005)' }
    PaymentActionRequest:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, description: 'Optimistic concurrency version' }
    PaymentConfirmRequest:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer }
        confirmSinglePerson: { type: boolean, description: 'Required for a maker self-confirm in single-person mode' }
    PaymentVoidRequest:
      type: object
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer }
        reason: { type: string }
    PaymentConfirmResult:
      allOf:
        - $ref: '#/components/schemas/PaymentDoc'
        - type: object
          properties:
            settlementVoucher: { $ref: '#/components/schemas/Voucher' }
...
    TrialBalanceRow:
      type: object
      required: [accountCode, accountName, openingDebit, openingCredit, periodDebit, periodCredit, closingDebit, closingCredit]
      properties:
        accountCode: { type: string }
        accountName: { type: string }
        openingDebit: { type: string }
        openingCredit: { type: string }
        periodDebit: { type: string }
        periodCredit: { type: string }
        closingDebit: { type: string }
        closingCredit: { type: string }
    TrialBalanceTotals:
      type: object
      required: [openingDebit, openingCredit, periodDebit, periodCredit, closingDebit, closingCredit]
      properties:
        openingDebit: { type: string }
        openingCredit: { type: string }
        periodDebit: { type: string }
        periodCredit: { type: string }
        closingDebit: { type: string }
        closingCredit: { type: string }
    TrialBalance:
      type: object
      required: [rows, totals, balanced]
      properties:
        rows: { type: array, items: { $ref: '#/components/schemas/TrialBalanceRow' } }
        totals: { $ref: '#/components/schemas/TrialBalanceTotals' }
        balanced:
          type: object
          properties:
            opening: { type: boolean }
            period: { type: boolean }
            closing: { type: boolean }
    LedgerBalance:
      type: object
      properties:
        debit: { type: string }
        credit: { type: string }
        balance: { type: string }
        balanceDir: { type: string, enum: ['借', '贷', '平'] }
...
    AccountLedger:
      type: object
      required: [accountCode, accountName, opening, rows, closing]
      properties:
        accountCode: { type: string }
        accountName: { type: string }
        opening: { $ref: '#/components/schemas/LedgerBalance' }
        closing: { $ref: '#/components/schemas/LedgerBalance' }
        rows:
          type: array
          items:
            type: object
            properties:
              date: { type: string, format: date }
              voucherNo: { type: string }
              voucherId: { type: string, format: uuid }
              summary: { type: string }
              debit: { type: string }
              credit: { type: string }
              balance: { type: string }
              balanceDir: { type: string, enum: ['借', '贷', '平'] }